==> 3. BaselineModels.ipynb <==
from google.colab import drive
drive.mount('/content/drive')

cd /content/drive/MyDrive/690F

import string
import numpy as np
import pandas as pd
from sklearn.model_selection import train_test_split
from sklearn.feature_extraction.text import CountVectorizer
from sklearn.metrics import accuracy_score, classification_report, confusion_matrix
from sklearn.feature_extraction.text import TfidfVectorizer
from sklearn.model_selection import GridSearchCV
from sklearn.pipeline import Pipeline
from sklearn.preprocessing import StandardScaler
from sklearn.compose import ColumnTransformer
from sklearn.linear_model import LogisticRegression
from sklearn.naive_bayes import MultinomialNB
import pickle
from sklearn.model_selection import train_test_split

train_df = pd.read_csv("final_processed_train_data.csv")

train_df.shape

train_df.columns
train_df.dropna(inplace = True)
to_show = ['id', 'target',
       'comment_text', 'asian', 'atheist', 'bisexual', 'black', 'buddhist',
       'christian', 'female', 'heterosexual', 'hindu',
       'homosexual_gay_or_lesbian', 'intellectual_or_learning_disability',
       'jewish', 'latino', 'male', 'muslim', 'other_disability',
       'other_gender', 'other_race_or_ethnicity', 'other_religion',
       'other_sexual_orientation', 'physical_disability',
       'psychiatric_or_mental_illness', 'transgender', 'white']
train_df.head().loc[:, to_show]

# # train_df = train_df_whole.sample(n=5000, random_state=1)
# train_df = pd.read_csv("processed_train_df.csv")
# print(len(train_df))

train_df = train_df[train_df['target'] >= 0.20]

# the number of non toxic comments was more than 10 times the number of toxic comments, we have subsampled it to the following distribution
print("Number of non toxic comments: "+str(len(train_df[train_df["target"] < 0.5])))
print("Number of toxic comments: "+str(len(train_df[train_df["target"] >= 0.5])))

train_df["prot_word_count"][train_df["prot_word_count"] > 0] = 1

print("Number of comments with bias inducing words not present: " + str(len(train_df[train_df["prot_word_count"] == 0])))
print("Number of comments with bias inducing words present: " + str(len(train_df[train_df["prot_word_count"] == 1])))

X = train_df[["comment_text", "prot_word_count", "edited_comment_text"]]
y = np.array(train_df['target'].apply(lambda x: x > .5).astype('int'))

print(len(X))

print(len(y))

# Baseline models - Learning with Z

from sklearn.feature_extraction.text import CountVectorizer
vectorizer = CountVectorizer()
Xvec = vectorizer.fit_transform(X['comment_text'])
X_vect = pd.DataFrame(Xvec.todense(), columns=vectorizer.get_feature_names_out())

X_vect

prot_attr = ['transgendered', 'sex', 'bigotry', 'racist', 'muslim', 'racism', 'priest', 'black', 'sexual', 'latino', 'man', 'islam', 'gay', 'men', 'straight', 'transgender', 'homosexual', 'jewish', 'catholic', 'heterosexual', 'women', 'church', 'religious', 'military', 'asian', 'hindu', 'buddhist', 'religion', 'atheist', 'white', 'chinese', 'male', 'christian', 'race', 'bisexual', 'jew', 'woman']

word_counts = {}
cols_to_drop = []
for i in list(X_vect.columns):
  X_vect[i][X_vect[i] > 0] = 1
  col_total = X_vect[i].sum()
  word_counts[i] = col_total
  if col_total < 11:
    cols_to_drop.append(i)

print("colstodrop",len(cols_to_drop))
print("orig_cols",len(X_vect.columns))
X_vect.drop(cols_to_drop, axis=1, inplace = True)
print("new_cols",len(X_vect.columns))

print(len(X_vect))

train_df['result'] = train_df['target'].apply(lambda x: x > .5).astype('int')

train_df.to_csv("train_df.csv")

X_train, X_test, y_train, y_test = train_test_split(X_vect, y, test_size=.4, stratify = y)

X_vect.to_csv("X_vect.csv")

from sklearn.linear_model import LogisticRegression
# Fitting a simple Logistic Regression on tf-idf
lr = LogisticRegression()
lr.fit(X_train, y_train)

print("Accuracy: "+str(lr.score(X_test, y_test)))
y_preds = lr.predict(X_test)


coeffs_ml_with_z = pd.DataFrame({"Feature" : X_vect.columns.tolist(),"Coefficients" : lr.coef_[0]})

prot_attr = ['transgendered', 'sex', 'bigotry', 'racist', 'muslim', 'racism', 'priest', 'black', 'sexual', 'latino', 'man', 'islam', 'gay', 'men', 'straight', 'transgender', 'homosexual', 'jewish', 'catholic', 'heterosexual', 'women', 'church', 'religious', 'military', 'asian', 'hindu', 'buddhist', 'religion', 'atheist', 'white', 'chinese', 'male', 'christian', 'race', 'bisexual', 'jew', 'woman']

coeffs_ml_with_z[coeffs_ml_with_z['Feature'].isin(prot_attr)]

res = coeffs_ml_with_z[coeffs_ml_with_z['Feature'].isin(prot_attr)]
res.to_csv("coeffs_baseline_lr.csv")

groups = {
'religion' : ['religion', 'religious', 'atheist', 'buddhist', 'christian', 'hindu', 'jewish', 'muslim', 'catholic', 'church', 'islam', 'jew', 'priest'],
'race' : ['asian', 'black', 'latino', 'white', 'chinese', 'race', 'racism'],
'gender' : ['female', 'male', 'transgender', 'man', 'men', 'trangendered', 'woman', 'women'],
'sexual_identity' : ['bisexual', 'heterosexual', 'gay', 'homosexual', 'sex', 'sexual', 'straight' ],
'misc' : ['bigotry', 'military']
}

avg_coeffs = {}

for group in groups:
    avg_coeff = np.mean(np.array(coeffs_ml_with_z[coeffs_ml_with_z['Feature'].isin(groups[group])]['Coefficients']))
    avg_coeffs[group] = avg_coeff

avg_coeffs
