==> 5. MIM model.ipynb <==
# FaX AI with SHAP and AIF360 Data

This example presents how we can use the FaX AI, [AIF360](https://github.com/Trusted-AI/AIF360), and [SHAP](https://github.com/slundberg/shap) libraries together for marrying fairness with explainability.  

We use the COMPAS dataset (using the AIF360 loading methods) and how to train and measure our methods, the methods implemented in the AIF360 library, and sklearn supervised learning methods.

%matplotlib inline
%load_ext autoreload
%autoreload 2
import matplotlib.pyplot as plt
import numpy as np
import pandas as pd
# import tensorflow.compat.v1 as tf
# tf.disable_eager_execution()
# tf.logging.set_verbosity(tf.logging.ERROR)

from sklearn.compose import make_column_transformer
from sklearn.linear_model import LogisticRegression
from sklearn.metrics import accuracy_score
from sklearn.model_selection import GridSearchCV, train_test_split
from sklearn.preprocessing import OneHotEncoder, StandardScaler
from sklearn.ensemble import RandomForestClassifier

# from aif360.sklearn.preprocessing import ReweighingMeta
from aif360.sklearn.inprocessing import AdversarialDebiasing, ExponentiatedGradientReduction, GridSearchReduction
from aif360.sklearn.postprocessing import CalibratedEqualizedOdds, PostProcessingMeta
from aif360.sklearn.datasets import fetch_adult, fetch_compas, fetch_german
from aif360.sklearn.metrics import disparate_impact_ratio, average_odds_error, generalized_fpr
from aif360.sklearn.metrics import generalized_fnr, difference, statistical_parity_difference,equal_opportunity_difference
import matplotlib.pyplot as plt

import shap
import FaX_methods
import aif360_utils

### Training MIM from FaX AI

data = pd.read_csv("X_vect.csv")
data.head()

len(data)

train_df = pd.read_csv("train_df.csv")

len(train_df)

data['prot_word_count'] = train_df['prot_word_count']

data['target'] = train_df['target']

# the number of non toxic comments was more than 10 times the number of toxic comments, we have subsampled it to the following distribution
print("Number of non toxic comments: "+str(len(data[data["target"] < 0.5])))
print("Number of toxic comments: "+str(len(data[data["target"] >= 0.5])))

sample = data.sample(66000)

# the number of non toxic comments was more than 10 times the number of toxic comments, we have subsampled it to the following distribution
print("Number of non toxic comments: "+str(len(sample[sample["target"] < 0.5])))
print("Number of toxic comments: "+str(len(sample[sample["target"] >= 0.5])))

y = sample['target'].apply(lambda x: x > .5).astype('int')

y

sample.columns

sample.drop(['Unnamed: 0', 'target'], axis = 1, inplace = True)

sample.columns

prot_attr = ['gay', 'black', 'islam', 'christian', 'white', 'muslim', 'homosexual', 'chinese', 'bisexual', 'jew', 'catholic', 'transgender', 'latino', 'hindu']

from sklearn.model_selection import train_test_split
X_train, X_test, y_train, y_test = train_test_split(sample, y, test_size = 0.3, stratify = y)
print(len(X_train))
print(len(X_test))

X_train.columns

X_train_dropped = X_train.drop(["prot_word_count"], axis = 1)
X_test_dropped = X_test.drop(["prot_word_count"], axis = 1)
mim_aif = aif360_utils.FaXAIF(X_train_dropped, y_train, prot_attr, model_type = 'MIM')

#predict the test data using the model
# pred_prob = mim_aif.predict_proba(X_test_dropped)
# print(pred_prob)
pred = mim_aif.predict(X_test_dropped)
print(pred)

from sklearn.metrics import accuracy_score
print(accuracy_score(y_test, pred))

from sklearn.linear_model import LogisticRegression
clf = LogisticRegression(random_state=0, max_iter = 400).fit(X_train_dropped, y_train)
pred_baseline = clf.predict(X_test_dropped)
print(accuracy_score(y_test, pred_baseline))

X_test["y_test"] = y_test
X_test["y_preds"] = pred
X_test["y_preds_baseline"] = pred_baseline

not_toxic = X_test[X_test["y_test"] == 0]
not_toxic_having_prot_attr = not_toxic[not_toxic["prot_word_count"] == 1]
print("Fair Model - % of non toxic comments in test having prot_attr and have been wrongly classified as toxic: "+str(len(not_toxic_having_prot_attr[not_toxic_having_prot_attr["y_preds"] == 1])/len(not_toxic_having_prot_attr)*100))
print("Baseline Model - % of non toxic comments in test having prot_attr and have been wrongly classified as toxic: "+str(len(not_toxic_having_prot_attr[not_toxic_having_prot_attr["y_preds_baseline"] == 1])/len(not_toxic_having_prot_attr)*100))

# Finding MDE

XZ = np.hstack((mim_aif.X, mim_aif.Z))
mde_mim = FaX_methods.MDE_ind(XZ)
inf_mim = mde_mim.influence(mim_aif.model, mim_aif.X)
cols = list(X_train_compas.columns)[1:] +[list(X_train_compas.columns)[0]] 
res = {cols[i]: inf_mim[i] for i in range(len(inf_mim))}
print(res)

for i in prot_attr:
    print(i)
    print(res[i])
